HSEARCH-1841 moved all things related to DeleteByQuery to SPI
