+++ engine/src/main/java/org/hibernate/search/backend/impl/lucene/NRTWorkspaceImpl.java
@@ -16,7 +16,6 @@ import java.util.concurrent.locks.ReentrantLock;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.hibernate.search.backend.AddLuceneWork;
-import org.hibernate.search.backend.DeleteByQueryLuceneWork;
 import org.hibernate.search.backend.DeleteLuceneWork;
 import org.hibernate.search.backend.FlushLuceneWork;
 import org.hibernate.search.backend.IndexWorkVisitor;
@@ -25,6 +24,7 @@ import org.hibernate.search.backend.OptimizeLuceneWork;
 import org.hibernate.search.backend.PurgeAllLuceneWork;
 import org.hibernate.search.backend.UpdateLuceneWork;
 import org.hibernate.search.backend.impl.CommitPolicy;
+import org.hibernate.search.backend.spi.DeleteByQueryLuceneWork;
 import org.hibernate.search.exception.AssertionFailure;
 import org.hibernate.search.indexes.spi.DirectoryBasedIndexManager;
 import org.hibernate.search.indexes.spi.DirectoryBasedReaderProvider;
