+++ engine/src/main/java/org/hibernate/search/backend/spi/DeleteByQueryWork.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.search.backend.spi;
 
-import org.hibernate.search.backend.DeletionQuery;
 import org.hibernate.search.backend.impl.DeleteByQuerySupport;
 
 /**
