+++ engine/src/test/java/org/hibernate/search/test/backend/DeleteByQueryTest.java
@@ -19,10 +19,10 @@ import org.hibernate.search.annotations.Field;
 import org.hibernate.search.annotations.Index;
 import org.hibernate.search.annotations.Indexed;
 import org.hibernate.search.annotations.Store;
-import org.hibernate.search.backend.DeletionQuery;
-import org.hibernate.search.backend.SingularTermDeletionQuery;
 import org.hibernate.search.backend.impl.DeleteByQuerySupport;
 import org.hibernate.search.backend.spi.DeleteByQueryWork;
+import org.hibernate.search.backend.spi.DeletionQuery;
+import org.hibernate.search.backend.spi.SingularTermDeletionQuery;
 import org.hibernate.search.backend.spi.Work;
 import org.hibernate.search.backend.spi.WorkType;
 import org.hibernate.search.backend.spi.Worker;
