+++ integrationtest/jms/src/test/java/org/hibernate/search/test/jms/master/JMSMasterTest.java
@@ -37,9 +37,9 @@ import org.hibernate.search.FullTextSession;
 import org.hibernate.search.engine.ProjectionConstants;
 import org.hibernate.search.Search;
 import org.hibernate.search.backend.AddLuceneWork;
-import org.hibernate.search.backend.DeleteByQueryLuceneWork;
 import org.hibernate.search.backend.LuceneWork;
-import org.hibernate.search.backend.SingularTermDeletionQuery;
+import org.hibernate.search.backend.spi.DeleteByQueryLuceneWork;
+import org.hibernate.search.backend.spi.SingularTermDeletionQuery;
 import org.hibernate.search.indexes.spi.IndexManager;
 import org.hibernate.search.query.engine.spi.HSQuery;
 import org.hibernate.search.test.SearchTestBase;
